+++ R/mutators.R
@@ -33,47 +33,28 @@ has_extra <- function(x) {
   x
 }
 # Groups =======================================================================
-#' @export
-#' @rdname groups
-#' @aliases get_groups,MultivariateAnalysis-method
-setMethod(
-  f = "get_groups",
-  signature = "MultivariateAnalysis",
-  definition = function(x, margin = 1) {
-    margin <- margin[[1L]]
-    if (margin == 1) grp <- x@rows@groups
-    if (margin == 2) grp <- x@columns@groups
-    grp
-  }
-)
+get_groups <- function(x, margin = 1) {
+  margin <- margin[[1L]]
+  if (margin == 1) grp <- x@rows@groups
+  if (margin == 2) grp <- x@columns@groups
+  grp
+}
 
-#' @export
-#' @rdname groups
-#' @aliases set_groups,MultivariateAnalysis-method
-setReplaceMethod(
-  f = "set_groups",
-  signature = "MultivariateAnalysis",
-  definition = function(x) {
-    if (is.null(value)) value <- character(0)
-    x@rows@groups <- value
-    methods::validObject(x)
-    x
-  }
-)
+`set_groups<-` <- function(x, margin = 1, value) {
+  if (is.null(value)) value <- character(0)
+  margin <- margin[[1L]]
+  if (margin == 1) x@rows@groups <- value
+  if (margin == 2) x@columns@groups <- value
+  methods::validObject(x)
+  x
+}
 
-#' @export
-#' @rdname groups
-#' @aliases has_groups,MultivariateAnalysis-method
-setMethod(
-  f = "has_groups",
-  signature = "MultivariateAnalysis",
-  definition = function(x, margin = 1) {
-    margin <- margin[[1L]]
-    if (margin == 1) grp <- x@rows@groups
-    if (margin == 2) grp <- x@columns@groups
-    length(grp) > 0
-  }
-)
+has_groups <- function(x, margin = 1) {
+  margin <- margin[[1L]]
+  if (margin == 1) grp <- x@rows@groups
+  if (margin == 2) grp <- x@columns@groups
+  length(grp) > 0
+}
 
 # Dimensions ===================================================================
 #' @export