+++ NEWS.md
@@ -1,4 +1,9 @@
 # dimensio 0.8.1.9000
+## Enhancements
+* Allow to display extra qualitative and quantitative variables in `viz_rows()`/`viz_individuals()` and `viz_columns()`/`viz_variables()`.
+
+## Bugfixes & changes
+* Deprecate `hightlight` argument of `viz_rows()`/`viz_individuals()` and `viz_columns()`/`viz_variables()`.
 
 # dimensio 0.8.1
 ## Internals