+++ NEWS.md
@@ -1,4 +1,4 @@
-# dimensio 0.8.1.9000
+# dimensio 0.9.0
 ## New classes and methods
 * Add `pcoa()` to compute principal coordinates analysis.